fix: restore blog page interactivity
Fix broken category filter and subscription form; add state management and validation.

#VERCEL_SKIP
